refactor(cmd): host firearweave in cmd/firearweave/cli
