+++ cmd/firearweave/cli/setup.go
@@ -1,4 +1,4 @@
-package firearweave
+package cli
 
 import (
 	"fmt"
