+++ cmd/firearweave/cli/tools.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package firearweave
+package cli
 
 import "github.com/ChainSafe/firehose-arweave/tools"
