+++ go.mod
@@ -4,7 +4,6 @@ go 1.16
 
 require (
 	github.com/ShinyTrinkets/overseer v0.3.0
-	github.com/abourget/llerrgroup v0.2.0 // indirect
 	github.com/dvsekhvalnov/jose2go v1.5.0
 	github.com/golang/protobuf v1.5.2
 	github.com/spf13/cobra v1.2.1
@@ -19,8 +18,8 @@ require (
 	github.com/streamingfast/dmetering v0.0.0-20220307162406-37261b4b3de9
 	github.com/streamingfast/dmetrics v0.0.0-20220307162521-2389094ab4a1
 	github.com/streamingfast/dstore v0.1.1-0.20220315134935-980696943a79
-	github.com/streamingfast/dtracing v0.0.0-20220305214756-b5c0e8699839 // indirect
 	github.com/streamingfast/firehose v0.1.1-0.20220307162825-0f43305f7436
+	github.com/streamingfast/firehose-arweave v0.0.0-00010101000000-000000000000
 	github.com/streamingfast/logging v0.0.0-20220304214715-bc750a74b424
 	github.com/streamingfast/merger v0.0.3-0.20220318152213-9ab5185b44e8
 	github.com/streamingfast/node-manager v0.0.2-0.20220405231057-da9a898ad97e
@@ -34,3 +33,5 @@ require (
 )
 
 replace github.com/ShinyTrinkets/overseer => github.com/streamingfast/overseer v0.2.1-0.20210326144022-ee491780e3ef
+
+replace github.com/streamingfast/firehose-arweave => /home/clearloop/firehose-arweave
